rename runtime config
