+++ src/js/wrg-header.js
@@ -1,5 +1,5 @@
 import { html, css, LitElement } from 'lit';
-import config from '../wrg-runtime-config.json';
+import config from '.././wrg-config.json.json';
 
 customElements.define(
   'wrg-header',
