+++ src/runtimeConfig.json.njk
@@ -1,5 +1,5 @@
 ---
-permalink: wrg-runtime-config.json
+permalink: ./wrg-config.json.json
 eleventyExcludeFromCollections: true
 ---
 {
